+++ CHANGELOG.md
@@ -512,3 +512,4 @@
 * Remove `Bindable` typeclass
 * Rename experimental `Bindable_` to `Bindable`
 * Use experimental `Extendable_` as superclass for `Comonad`
+* Remove `Extendable` typeclass